Add DebugInterpreter to house heavy-lifting
- Move breakpoints/watch expression management/command interpretation logic
  out of lib/commands/debug and into new lib/debug/interpreter module

- Define async setupDebugger function as inner function to l/c/d's run

- Use run() purely to start interpreter and pass termination callback

- Setup/manage DebugPrinter inside DebugInterpreter
