+++ packages/truffle-core/lib/commands/debug.js
@@ -17,12 +17,16 @@ const command = {
     ]
   },
   run: function(options, done) {
-    const path = require("path");
+    command
+      .setupDebugger(options)
+      .then(interpreter => interpreter.start(done))
+      .catch(done);
+  },
+  setupDebugger: async function(options) {
     const debugModule = require("debug");
     const debug = debugModule("lib:commands:debug");
     const util = require("util");
     const BN = require("bn.js");
-    const analytics = require("../services/analytics");
     const ora = require("ora");
 
     // add custom inspect options for BNs
@@ -35,622 +39,84 @@ const command = {
     const Debugger = require("truffle-debugger");
     const DebugUtils = require("truffle-debug-utils");
     const Environment = require("../environment");
-    const ReplManager = require("../repl");
-    const selectors = require("truffle-debugger").selectors;
-    const { DebugPrinter } = require("../debug");
-
-    // Debugger Session properties
-    const trace = selectors.trace;
-    const solidity = selectors.solidity;
-    const controller = selectors.controller;
+    const { DebugInterpreter } = require("../debug");
 
     const config = Config.detect(options);
 
     config.logger.log("Starting Truffle Debugger...");
 
-    Environment.detect(config)
-      .then(() => {
-        const txHash = config._[0]; //may be undefined
+    await Environment.detect(config);
 
-        let lastCommand = "n";
-        const enabledExpressions = new Set();
-        let startSpinner; //apologies for the use of a global variable here
-        let compileSpinner; //and here
+    const txHash = config._[0]; //may be undefined
 
-        let compilePromise = new Promise(function(accept, reject) {
-          //we need to set up a config object for the compiler.
-          //it's the same as the existing config, but we turn on quiet.
-          //unfortunately, we don't have Babel here, so cloning is annoying.
-          let compileConfig = Object.assign(
-            {},
-            ...Object.entries(config).map(([key, value]) => ({ [key]: value }))
-          ); //clone
-          compileConfig.quiet = true;
+    let startSpinner; //apologies for the use of a global variable here
+    let compileSpinner; //and here
 
-          compileSpinner = ora("Compiling your contracts...").start();
+    const compilation = await new Promise(function(accept, reject) {
+      //we need to set up a config object for the compiler.
+      //it's the same as the existing config, but we turn on quiet.
+      //unfortunately, we don't have Babel here, so cloning is annoying.
+      let compileConfig = Object.assign(
+        {},
+        ...Object.entries(config).map(([key, value]) => ({ [key]: value }))
+      ); //clone
+      compileConfig.quiet = true;
 
-          compile.all(compileConfig, function(err, contracts, files) {
-            if (err) {
-              return reject(err);
-            }
+      compileSpinner = ora("Compiling your contracts...").start();
 
-            return accept({
-              contracts: contracts,
-              files: files
-            });
-          });
+      compile.all(compileConfig, function(err, contracts, files) {
+        if (err) {
+          return reject(err);
+        }
+
+        return accept({
+          contracts: contracts,
+          files: files
         });
-
-        const sessionPromise = compilePromise
-          .then(function(result) {
-            compileSpinner.succeed();
-            let startMessage = DebugUtils.formatStartMessage(
-              txHash !== undefined
-            );
-            startSpinner = ora(startMessage).start();
-
-            let debuggerOptions = {
-              provider: config.provider,
-              files: result.files,
-              contracts: Object.keys(result.contracts).map(function(name) {
-                const contract = result.contracts[name];
-                return {
-                  contractName: contract.contractName || contract.contract_name,
-                  source: contract.source,
-                  sourcePath: contract.sourcePath,
-                  ast: contract.ast,
-                  binary: contract.binary || contract.bytecode,
-                  sourceMap: contract.sourceMap,
-                  deployedBinary:
-                    contract.deployedBinary || contract.deployedBytecode,
-                  deployedSourceMap: contract.deployedSourceMap,
-                  compiler: contract.compiler,
-                  abi: contract.abi
-                };
-              })
-            };
-
-            return txHash !== undefined
-              ? Debugger.forTx(txHash, debuggerOptions)
-              : Debugger.forProject(debuggerOptions);
-          })
-          .then(function(bugger) {
-            debug("about to connect");
-            return bugger.connect();
-          })
-          .catch(done);
-
-        sessionPromise
-          .then(async function(session) {
-            const printer = new DebugPrinter(config, session);
-
-            function watchExpressionAnalytics(raw) {
-              if (raw.includes("!<")) {
-                //don't send analytics for watch expressions involving selectors
-                return;
-              }
-              let expression = raw.trim();
-              //legal Solidity identifiers (= legal JS identifiers)
-              let identifierRegex = /^[a-zA-Z_$][a-zA-Z_$0-9]*$/;
-              let isVariable = expression.match(identifierRegex) !== null;
-              analytics.send({
-                command: "debug: watch expression",
-                args: { isVariable }
-              });
-            }
-
-            async function setOrClearBreakpoint(args, setOrClear) {
-              //setOrClear: true for set, false for clear
-              const currentLocation = session.view(controller.current.location);
-              const breakpoints = session.view(controller.breakpoints);
-
-              const currentNode = currentLocation.node
-                ? currentLocation.node.id
-                : null;
-              const currentLine = currentLocation.sourceRange
-                ? currentLocation.sourceRange.lines.start.line
-                : null;
-              const currentSourceId = currentLocation.source
-                ? currentLocation.source.id
-                : null;
-
-              let breakpoint = {};
-
-              if (args.length === 0) {
-                //no arguments, want currrent node
-                debug("node case");
-                if (currentNode === null) {
-                  config.logger.log("Cannot determine current location.");
-                  return;
-                }
-                breakpoint.node = currentNode;
-                breakpoint.line = currentLine;
-                breakpoint.sourceId = currentSourceId;
-              }
-
-              //the special case of "B all"
-              else if (args[0] === "all") {
-                if (setOrClear) {
-                  // only "B all" is legal, not "b all"
-                  config.logger.log("Cannot add breakpoint everywhere.\n");
-                  return;
-                }
-                await session.removeAllBreakpoints();
-                config.logger.log("Removed all breakpoints.\n");
-                return;
-              }
-
-              //if the argument starts with a "+" or "-", we have a relative
-              //line number
-              else if (args[0][0] === "+" || args[0][0] === "-") {
-                debug("relative case");
-                if (currentLine === null) {
-                  config.logger.log("Cannot determine current location.");
-                  return;
-                }
-                let delta = parseInt(args[0], 10); //want an integer
-                debug("delta %d", delta);
-
-                if (isNaN(delta)) {
-                  config.logger.log("Offset must be an integer.\n");
-                  return;
-                }
-
-                breakpoint.sourceId = currentSourceId;
-                breakpoint.line = currentLine + delta;
-              }
-
-              //if it contains a colon, it's in the form source:line
-              else if (args[0].includes(":")) {
-                debug("source case");
-                let sourceArgs = args[0].split(":");
-                let sourceArg = sourceArgs[0];
-                let lineArg = sourceArgs[1];
-                debug("sourceArgs %O", sourceArgs);
-
-                //first let's get the line number as usual
-                let line = parseInt(lineArg, 10); //want an integer
-                if (isNaN(line)) {
-                  config.logger.log("Line number must be an integer.\n");
-                  return;
-                }
-
-                //search sources for given string
-                let sources = session.view(solidity.info.sources);
-
-                //we will indeed need the sources here, not just IDs
-                let matchingSources = Object.values(sources).filter(source =>
-                  source.sourcePath.includes(sourceArg)
-                );
-
-                if (matchingSources.length === 0) {
-                  config.logger.log(
-                    `No source file found matching ${sourceArg}.\n`
-                  );
-                  return;
-                } else if (matchingSources.length > 1) {
-                  config.logger.log(
-                    `Multiple source files found matching ${sourceArg}.  Which did you mean?`
-                  );
-                  matchingSources.forEach(source =>
-                    config.logger.log(source.sourcePath)
-                  );
-                  config.logger.log("");
-                  return;
-                }
-
-                //otherwise, we found it!
-                breakpoint.sourceId = matchingSources[0].id;
-                breakpoint.line = line - 1; //adjust for zero-indexing!
-              }
-
-              //otherwise, it's a simple line number
-              else {
-                debug("absolute case");
-                if (currentSourceId === null) {
-                  config.logger.log("Cannot determine current file.");
-                  return;
-                }
-                let line = parseInt(args[0], 10); //want an integer
-                debug("line %d", line);
-
-                if (isNaN(line)) {
-                  config.logger.log("Line number must be an integer.\n");
-                  return;
-                }
-
-                breakpoint.sourceId = currentSourceId;
-                breakpoint.line = line - 1; //adjust for zero-indexing!
-              }
-
-              //OK, we've constructed the breakpoint!  But if we're adding, we'll
-              //want to adjust to make sure we don't set it on an empty line or
-              //anything like that
-              if (setOrClear) {
-                let resolver = session.view(controller.breakpoints.resolver);
-                breakpoint = resolver(breakpoint);
-                //of course, this might result in finding that there's nowhere to
-                //add it after that point
-                if (breakpoint === null) {
-                  config.logger.log(
-                    "Nowhere to add breakpoint at or beyond that location.\n"
-                  );
-                  return;
-                }
-              }
-
-              //having constructed and adjusted breakpoint, here's now a
-              //user-readable message describing its location
-              let sourceNames = Object.assign(
-                {},
-                ...Object.values(session.view(solidity.info.sources)).map(
-                  ({ id, sourcePath }) => ({
-                    [id]: path.basename(sourcePath)
-                  })
-                )
-              );
-              let locationMessage = DebugUtils.formatBreakpointLocation(
-                breakpoint,
-                true,
-                currentSourceId,
-                sourceNames
-              );
-
-              //one last check -- does this breakpoint already exist?
-              let alreadyExists =
-                breakpoints.filter(
-                  existingBreakpoint =>
-                    existingBreakpoint.sourceId === breakpoint.sourceId &&
-                    existingBreakpoint.line === breakpoint.line &&
-                    existingBreakpoint.node === breakpoint.node //may be undefined
-                ).length > 0;
-
-              //NOTE: in the "set breakpoint" case, the above check is somewhat
-              //redundant, as we're going to check again when we actually make the
-              //call to add or remove the breakpoint!  But we need to check here so
-              //that we can display the appropriate message.  Hopefully we can find
-              //some way to avoid this redundant check in the future.
-
-              //if it already exists and is being set, or doesn't and is being
-              //cleared, report back that we can't do that
-              if (setOrClear === alreadyExists) {
-                if (setOrClear) {
-                  config.logger.log(
-                    `Breakpoint at ${locationMessage} already exists.\n`
-                  );
-                  return;
-                } else {
-                  config.logger.log(
-                    `No breakpoint at ${locationMessage} to remove.\n`
-                  );
-                  return;
-                }
-              }
-
-              //finally, if we've reached this point, do it!
-              //also report back to the user on what happened
-              if (setOrClear) {
-                await session.addBreakpoint(breakpoint);
-                config.logger.log(`Breakpoint added at ${locationMessage}.\n`);
-              } else {
-                await session.removeBreakpoint(breakpoint);
-                config.logger.log(
-                  `Breakpoint removed at ${locationMessage}.\n`
-                );
-              }
-              return;
-            }
-
-            function setPrompt(prompt) {
-              repl.activate.bind(repl)({
-                prompt,
-                context: {},
-                //this argument only *adds* things, so it's safe to set it to {}
-                ignoreUndefined: true
-                //set to true because it's set to true below :P
-              });
-            }
-
-            async function interpreter(cmd) {
-              cmd = cmd.trim();
-              let cmdArgs, splitArgs;
-              debug("cmd %s", cmd);
-
-              if (cmd === ".exit") {
-                cmd = "q";
-              }
-
-              //split arguments for commands that want that; split on runs of spaces
-              splitArgs = cmd
-                .trim()
-                .split(/ +/)
-                .slice(1);
-              debug("splitArgs %O", splitArgs);
-
-              //warning: this bit *alters* cmd!
-              if (cmd.length > 0) {
-                cmdArgs = cmd.slice(1).trim();
-                cmd = cmd[0];
-              }
-
-              if (cmd === "") {
-                cmd = lastCommand;
-                cmdArgs = "";
-                splitArgs = [];
-              }
-
-              //quit if that's what we were given
-              if (cmd === "q") {
-                return await util.promisify(repl.stop.bind(repl))();
-              }
-
-              let alreadyFinished = session.view(trace.finishedOrUnloaded);
-              let loadFailed = false;
-
-              // If not finished, perform commands that require state changes
-              // (other than quitting or resetting)
-              if (!alreadyFinished) {
-                let stepSpinner = ora("Stepping...").start();
-                switch (cmd) {
-                  case "o":
-                    await session.stepOver();
-                    break;
-                  case "i":
-                    await session.stepInto();
-                    break;
-                  case "u":
-                    await session.stepOut();
-                    break;
-                  case "n":
-                    await session.stepNext();
-                    break;
-                  case ";":
-                    //two cases -- parameterized and unparameterized
-                    if (cmdArgs !== "") {
-                      let count = parseInt(cmdArgs, 10);
-                      debug("cmdArgs=%s", cmdArgs);
-                      if (isNaN(count)) {
-                        config.logger.log(
-                          "Number of steps must be an integer."
-                        );
-                        break;
-                      }
-                      await session.advance(count);
-                    } else {
-                      await session.advance();
-                    }
-                    break;
-                  case "c":
-                    await session.continueUntilBreakpoint();
-                    break;
-                }
-                stepSpinner.stop();
-              } //otherwise, inform the user we can't do that
-              else {
-                switch (cmd) {
-                  case "o":
-                  case "i":
-                  case "u":
-                  case "n":
-                  case "c":
-                    //are we "finished" because we've reached the end, or because
-                    //nothing is loaded?
-                    if (session.view(selectors.session.status.loaded)) {
-                      config.logger.log(
-                        "Transaction has halted; cannot advance."
-                      );
-                      config.logger.log("");
-                    } else {
-                      config.logger.log("No transaction loaded.");
-                      config.logger.log("");
-                    }
-                }
-              }
-              if (cmd === "r") {
-                //reset if given the reset command
-                //(but not if nothing is loaded)
-                if (session.view(selectors.session.status.loaded)) {
-                  await session.reset();
-                } else {
-                  config.logger.log("No transaction loaded.");
-                  config.logger.log("");
-                }
-              }
-              if (cmd === "t") {
-                if (!session.view(selectors.session.status.loaded)) {
-                  let txSpinner = ora(
-                    DebugUtils.formatTransactionStartMessage()
-                  ).start();
-                  await session.load(cmdArgs);
-                  //if load succeeded
-                  if (session.view(selectors.session.status.success)) {
-                    txSpinner.succeed();
-                    //if successful, change prompt
-                    setPrompt(DebugUtils.formatPrompt(config.network, cmdArgs));
-                  } else {
-                    txSpinner.fail();
-                    loadFailed = true;
-                  }
-                } else {
-                  loadFailed = true;
-                  config.logger.log(
-                    "Please unload the current transaction before loading a new one."
-                  );
-                }
-              }
-              if (cmd === "T") {
-                if (session.view(selectors.session.status.loaded)) {
-                  await session.unload();
-                  config.logger.log("Transaction unloaded.");
-                  setPrompt(DebugUtils.formatPrompt(config.network));
-                } else {
-                  config.logger.log("No transaction to unload.");
-                  config.logger.log("");
-                }
-              }
-
-              // Check if execution has (just now) stopped.
-              if (session.view(trace.finished) && !alreadyFinished) {
-                config.logger.log("");
-                //check if transaction failed
-                if (
-                  !session.view(selectors.session.transaction.receipt).status
-                ) {
-                  config.logger.log("Transaction halted with a RUNTIME ERROR.");
-                  config.logger.log("");
-                  config.logger.log(
-                    "This is likely due to an intentional halting expression, like assert(), require() or revert(). It can also be due to out-of-gas exceptions. Please inspect your transaction parameters and contract code to determine the meaning of this error."
-                  );
-                } else {
-                  //case if transaction succeeded
-                  config.logger.log("Transaction completed successfully.");
-                }
-              }
-
-              // Perform post printing
-              // (we want to see if execution stopped before printing state).
-              switch (cmd) {
-                case "+":
-                  if (cmdArgs[0] === ":") {
-                    watchExpressionAnalytics(cmdArgs.substring(1));
-                  }
-                  enabledExpressions.add(cmdArgs);
-                  await printer.printWatchExpressionResult(cmdArgs);
-                  break;
-                case "-":
-                  enabledExpressions.delete(cmdArgs);
-                  break;
-                case "!":
-                  printer.printSelector(cmdArgs);
-                  break;
-                case "?":
-                  printer.printWatchExpressions(enabledExpressions);
-                  printer.printBreakpoints();
-                  break;
-                case "v":
-                  await printer.printVariables();
-                  break;
-                case ":":
-                  watchExpressionAnalytics(cmdArgs);
-                  printer.evalAndPrintExpression(cmdArgs);
-                  break;
-                case "b":
-                  await setOrClearBreakpoint(splitArgs, true);
-                  break;
-                case "B":
-                  await setOrClearBreakpoint(splitArgs, false);
-                  break;
-                case ";":
-                case "p":
-                  if (session.view(selectors.session.status.loaded)) {
-                    printer.printFile();
-                    printer.printInstruction();
-                    printer.printState();
-                  }
-                  await printer.printWatchExpressionsResults(
-                    enabledExpressions
-                  );
-                  break;
-                case "o":
-                case "i":
-                case "u":
-                case "n":
-                case "c":
-                  if (!session.view(trace.finishedOrUnloaded)) {
-                    if (!session.view(solidity.current.source).source) {
-                      printer.printInstruction();
-                    }
-                    printer.printFile();
-                    printer.printState();
-                  }
-                  await printer.printWatchExpressionsResults(
-                    enabledExpressions
-                  );
-                  break;
-                case "r":
-                  if (session.view(selectors.session.status.loaded)) {
-                    printer.printAddressesAffected();
-                    printer.printFile();
-                    printer.printState();
-                  }
-                  break;
-                case "t":
-                  if (!loadFailed) {
-                    printer.printAddressesAffected();
-                    printer.printFile();
-                    printer.printState();
-                  } else if (session.view(selectors.session.status.isError)) {
-                    let loadError = session.view(
-                      selectors.session.status.error
-                    );
-                    config.logger.log(loadError);
-                  }
-                  break;
-                case "T":
-                  //nothing to print
-                  break;
-                default:
-                  printer.printHelp();
-              }
-
-              if (
-                cmd !== "i" &&
-                cmd !== "u" &&
-                cmd !== "b" &&
-                cmd !== "B" &&
-                cmd !== "v" &&
-                cmd !== "h" &&
-                cmd !== "p" &&
-                cmd !== "?" &&
-                cmd !== "!" &&
-                cmd !== ":" &&
-                cmd !== "+" &&
-                cmd !== "r" &&
-                cmd !== "-" &&
-                cmd !== "t" &&
-                cmd !== "T"
-              ) {
-                lastCommand = cmd;
-              }
-            }
-
-            let prompt;
-
-            if (session.view(selectors.session.status.loaded)) {
-              startSpinner.succeed();
-              printer.printAddressesAffected();
-              printer.printHelp();
-              debug("Help printed");
-              printer.printFile();
-              debug("File printed");
-              printer.printState();
-              debug("State printed");
-              prompt = DebugUtils.formatPrompt(config.network, txHash);
-            } else {
-              if (session.view(selectors.session.status.isError)) {
-                startSpinner.fail();
-                config.logger.log(session.view(selectors.session.status.error));
-              } else {
-                startSpinner.succeed();
-              }
-              printer.printHelp();
-              prompt = DebugUtils.formatPrompt(config.network);
-            }
-
-            const repl = options.repl || new ReplManager(config);
-
-            repl.start({
-              prompt,
-              interpreter: util.callbackify(interpreter),
-              ignoreUndefined: true,
-              done: done
-            });
-          })
-          .catch(done);
-      })
-      .catch(error => {
-        done(error);
       });
+    });
+
+    compileSpinner.succeed();
+
+    let startMessage = DebugUtils.formatStartMessage(txHash !== undefined);
+    startSpinner = ora(startMessage).start();
+
+    let debuggerOptions = {
+      provider: config.provider,
+      files: compilation.files,
+      contracts: Object.keys(compilation.contracts).map(function(name) {
+        const contract = compilation.contracts[name];
+        return {
+          contractName: contract.contractName || contract.contract_name,
+          source: contract.source,
+          sourcePath: contract.sourcePath,
+          ast: contract.ast,
+          binary: contract.binary || contract.bytecode,
+          sourceMap: contract.sourceMap,
+          deployedBinary: contract.deployedBinary || contract.deployedBytecode,
+          deployedSourceMap: contract.deployedSourceMap,
+          compiler: contract.compiler,
+          abi: contract.abi
+        };
+      })
+    };
+
+    const bugger =
+      txHash !== undefined
+        ? await Debugger.forTx(txHash, debuggerOptions)
+        : await Debugger.forProject(debuggerOptions);
+
+    debug("about to connect");
+    const session = bugger.connect();
+
+    const interpreter = new DebugInterpreter(
+      config,
+      session,
+      startSpinner,
+      txHash
+    );
+
+    return interpreter;
   }
 };
